% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmtest.r
\name{computeDMStat}
\alias{computeDMStat}
\title{compute Delgado and Manteiga (2001) test statistic}
\usage{
computeDMStat(Y, X, Z, a = NA, ckertype = "gaussian", stat = "CvM")
}
\arguments{
\item{Y}{n-dim. vector containing the observations on the outcome}

\item{X}{matrix with n rows containing the observations on the scalar or vector X}

\item{Z}{matrix with n rows containing the observations on the scalar or vector Z}

\item{a}{vector of bandwidths, of the same dimension as there are columns in X, if unspecified, then the bandwidths are determined by cross-validation from nonparametric regression of Y on X}

\item{ckertype}{character string denoting the kernel function to be used, as in np package (default: "gaussian")}

\item{stat}{character string denoting the type of test statistic to be computed: Cramer-von-Mises ("CvM", default) or Kolmogorov-Smirnov ("KS")}
}
\value{
a list containing the following elements: 'teststat' value of the test statistic, 'epsilonhat' the residuals from a nonparametric regression from Y on X, 'Yhat' the predicted values from a nonparametric regression of Y on X, 'a' the bandwidth(s)
}
\description{
compute Delgado and Manteiga (2001) test statistic for testing the hypothesis H0: E[Y|X,Z] = E[Y|X]
}
\examples{
Y <- rnorm(100)
X <- rnorm(100)
Z <- rnorm(100)
computeDMStat(Y, X, Z, a=NA, ckertype="gaussian", stat="CvM")
}
\keyword{Delgado}
\keyword{Manteiga}
\keyword{error}
\keyword{measurement}
\keyword{significance}
\keyword{test}
