% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmtest.r
\name{computeTStat}
\alias{computeTStat}
\title{compute Delgado and Manteiga (2001) test statistic}
\usage{
computeTStat(Y, X, Z, a = NA, ckertype = "gaussian", stat = "CvM")
}
\arguments{
\item{Y}{n-dim. vector containing the observations on the outcome}

\item{X}{matrix with n rows containing the observations on the variable X}

\item{Z}{matrix with n rows containing the observations on the variable Z}

\item{a}{vector of bandwidths, of the same dimension as there are columns in X, if unspecified, then the bandwidths are determined by cross-validation from nonparametric regression of Y on X}

\item{ckertype}{character string denoting the kernel function to be used, as in np package (default: "gaussian")}

\item{stat}{character string denoting the type of test statistic to be computed: Cramer-von-Mises ("CvM", default) or Kolmogorov-Smirnov ("KS")}
}
\value{
(teststat, epsilonhat, Yhat) value of the test statistic, the residuals epsilonhat, and predicted values Yhat
}
\description{
compute Delgado and Manteiga (2001) test statistic for testing the hypothesis H0: E[Y|X,Z] = E[Y|X]
}
\examples{
computeTStat(Y, X, Z, a=NA, ckertype="gaussian", stat="CvM")
}
\keyword{Delgado}
\keyword{Manteiga}
\keyword{error}
\keyword{measurement}
\keyword{significance}
\keyword{test}

