% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmtest.r
\name{DMTest}
\alias{DMTest}
\title{perform the Delgado and Manteiga (2001) test}
\usage{
DMTest(
  Y,
  X,
  Z,
  size = 0.05,
  B = 100,
  a = NA,
  ckertype = "gaussian",
  stat = "CvM",
  indep = FALSE
)
}
\arguments{
\item{Y}{n-dim. vector containing the observations on the outcome}

\item{X}{matrix with n rows containing the observations on the scalar or vector X}

\item{Z}{matrix with n rows containing the observations on the scalar or vector Z}

\item{size}{scalar between 0 and 1, denoting the nominal size of the test (default: 0.05)}

\item{B}{integer denoting the number of bootstrap samples to be used (default: 100)}

\item{a}{vector of bandwidths, of the same dimension as there are columns in X, if unspecified, then the bandwidths are determined by cross-validation from nonparametric regression of Y on X}

\item{ckertype}{character string denoting the kernel function to be used, as in np package (default: "gaussian")}

\item{stat}{character string denoting the type of test statistic to be computed: Cramer-von-Mises ("CvM", default) or Kolmogorov-Smirnov ("KS")}

\item{indep}{logical; FALSE means that conditional mean independence is tested, otherwise conditional independence}
}
\value{
a list containing the following elements: 'teststat' value of the test statistic, 'cv' bootstrap critical value, 'rej' a 1-0 indicator for whether the test rejects or not, 'pval' p-value, 'a' the bandwidth(s)
}
\description{
perform the Delgado and Manteiga (2001) test of the hypothesis of conditional mean independence (E[Y|X,Z] = E[Y|X]) or conditional independence (Y is independent of Z given X)
}
\examples{
Y <- rnorm(100)
X <- rnorm(100)
Z <- rnorm(100)
DMTest(Y, X, Z, size=0.05, B=100, a=NA, ckertype="gaussian", stat="CvM")
}
\keyword{Delgado}
\keyword{Manteiga}
\keyword{error}
\keyword{measurement}
\keyword{significance}
\keyword{test}
