% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmtest.r
\name{rMammen}
\alias{rMammen}
\title{draw a random sample from Mammen's two-point distribution}
\usage{
rMammen(n)
}
\arguments{
\item{n}{sample size}
}
\value{
V random sample from Mammen's two-point distribution
}
\description{
draw a random sample from Mammen's two-point distribution
}
\examples{
rMammen(n)
}
\keyword{Mammen}
\keyword{generator}
\keyword{random}
\keyword{sample}

